taskmanager/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.2.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
comm==0.2.3
cryptography==46.0.1
cssselect2==0.8.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.6
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-signature-pad==0.7.0
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.59.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
lxml==6.0.1
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
openai==1.107.1
oscrypto==1.3.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
yarg==0.1.9
zopfli==0.2.3.post1
